optimize the ci config and rename the urma api mock source name
