+++ mooncake-transfer-engine/src/transport/kunpeng_transport/CMakeLists.txt
@@ -1,4 +1,4 @@
-file(GLOB UB_SOURCES "*.cpp")
+file(GLOB_RECURSE UB_SOURCES "*.cpp")
 
 # Check if liburma.so exists
 find_library(URMA_LIBRARY urma PATHS /usr/lib64)
@@ -16,7 +16,7 @@ if (URMA_LIBRARY)
 else()
     # If liburma.so not found, we'll need to handle this differently
     # For now, just build without mock_urma_api.cpp
-    list(APPEND UB_SOURCES "urma/mock_urma_api.cpp")
+    list(APPEND UB_SOURCES "urma/mock_urma.cpp")
     add_library(ub_transport OBJECT ${UB_SOURCES})
     message(WARNING "Not Found liburma.so building ub_transport with Mock URMA library")
 endif()
